Improved stability with retry attempts
For each request to control a device, this script will make 3 attempts at getting an HTTP 200 status from the API for the devices information (before it only tried once).  Once it gets a 200 status, it will move on towards sending the singal to control the device 10 times (before it only tried once).  This makes the program significantly more likely to successfully control a device that it can't get feedback from, like an RF switch.  From my personal testing, this is much more reliable, but still not perfect.  Since RF devices are unable to provide feedback, it will never be possible to guarentee they get controlled.

Also fixed some bugs related to it attempting to print an `int` instead of a `String` in a debug message.

Also used better types for variables like `size_t` and usage of `unsigned`.