Update to use public ID's for users and devices
The microcontroller is now solely dependent on controlling and asking about users and devices based on the public ID's which are UUID's.  This adds a layer of security because it was previously using integer based ID's which are easy to guess and iterate through.  While still technically possible to guess UUID's, the odds are beyond astronomically small that it is not a worthwhile attack.